clean build logic
